backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# LLM & RAG
groq==0.11.0
llama-index==0.11.4
llama-index-llms-groq==0.2.0
llama-index-embeddings-huggingface==0.3.1

# Vector Store
faiss-cpu==1.8.0
numpy==1.26.4

# PDF Parsing
pymupdf==1.24.10
pytesseract==0.3.13
Pillow==10.4.0

# HTTP & Scraping
httpx==0.27.2
beautifulsoup4==4.12.3

# Database & Cache
sqlalchemy==2.0.35
asyncpg==0.29.0
redis==5.1.0

# Validation & Settings
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-dotenv==1.0.1

# Utilities
aiofiles==24.1.0
tenacity==9.0.0
loguru==0.7.2

# Dev & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2